// src/starshipsTypes.d.ts
export type StarshipsTypes = {
    name: string
    model: string
    manufacturer: string
    cost_in_credits: string
    length: string
    max_atmosphering_speed: string
    crew: string
    passengers: string
    cargo_capacity: string
    consumables: string
    hyperdrive_rating: string
    MGLT: string
    starship_class: string
    pilots: string[] // This assumes the pilot URLs are strings
    films: string[] // This assumes the film URLs are strings
    created: string
    edited: string
    url: string
}